Add lighting to illuminated objects
