+++ Assets/Controller Scripts/LightRenderController.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97133244c2db6c54d9a3fe2ad479f33d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
