+++ Assets/Controller Scripts/LightRenderController_Renderer.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a051c14168583cd4882c525b8b3d8c43
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
