+++ Assets/_ProceduralGeneration/RuleTile/LevelRuleTile.asset
@@ -45,7 +45,7 @@ MonoBehaviour:
     m_MaxAnimationSpeed: 1
     m_PerlinScale: 0.5
     m_Output: 0
-    m_ColliderType: 0
+    m_ColliderType: 1
     m_RandomTransform: 0
     m_Neighbors: 0100000001000000010000000100000001000000010000000100000002000000
     m_NeighborPositions:
